Refactor gui
